+++ src/lib/util/chdcd.c
@@ -995,7 +995,7 @@ chd_error chdcd_parse_cue(const char *tocfname, cdrom_toc &outtoc, chdcd_track_i
 chd_error chdcd_parse_toc(const char *tocfname, cdrom_toc &outtoc, chdcd_track_input_info &outinfo)
 {
 	FILE *infile;
-	size_t i, trknum;
+	int i, trknum;
 	static char token[512];
 	char tocftemp[512];
